+++ src/main/java/ru/yandex/practicum/filmorate/FilmorateApplication.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum;
+package ru.yandex.practicum.filmorate;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
