+++ src/main/java/ru/yandex/practicum/filmorate/exceptions/ValidationException.java
@@ -1,4 +1,5 @@
-package ru.yandex.practicum.exceptions;
+package ru.yandex.practicum.filmorate.exceptions;
+
 public class ValidationException extends RuntimeException {
     public ValidationException(String message) {
         super(message);
