+++ src/main/java/ru/yandex/practicum/filmorate/validannotation/FilmReleaseDateConstraint.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.validannotation;
+package ru.yandex.practicum.filmorate.validannotation;
 
 import javax.validation.Constraint;
 import javax.validation.Payload;
@@ -12,6 +12,8 @@ import java.lang.annotation.Target;
 @Constraint(validatedBy = ReleaseDateValidator.class)
 public @interface FilmReleaseDateConstraint {
     String message() default "Must be after 28.12.1895";
-    Class<?>[] groups() default {};
-    Class<?extends Payload>[] payload() default {};
+
+    Class <?>[] groups() default {};
+
+    Class <? extends Payload>[] payload() default {};
 }
