+++ src/main/java/ru/yandex/practicum/filmorate/validannotation/ReleaseDateValidator.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.validannotation;
+package ru.yandex.practicum.filmorate.validannotation;
 
 import javax.validation.ConstraintValidator;
 import javax.validation.ConstraintValidatorContext;
@@ -7,6 +7,7 @@ import java.time.LocalDate;
 public class ReleaseDateValidator implements ConstraintValidator<FilmReleaseDateConstraint, LocalDate> {
     @Override
     public boolean isValid(LocalDate releaseDate, ConstraintValidatorContext constraintValidatorContext) {
-        return releaseDate.equals(LocalDate.of(1895, 12, 28)) || releaseDate.isAfter(LocalDate.of(1895, 12, 28));
+        return releaseDate.equals(LocalDate.of(1895, 12, 28)) ||
+                releaseDate.isAfter(LocalDate.of(1895, 12, 28)) ;
     }
 }
