--- src/main/java/ru/yandex/practicum/model/Film.java
@@ -1,30 +1,0 @@
-package ru.yandex.practicum.model;
-
-import lombok.*;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import java.time.LocalDate;
-
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor(force = true)
-public class Film {
-    @EqualsAndHashCode.Exclude
-    private int id;
-    @NonNull
-    private String name;
-    @NonNull
-    private String description;
-    @NonNull
-    private LocalDate releaseDate;
-    @NonNull
-    private int duration;
-
-    public Film(@NonNull String name, @NonNull String description, @NonNull LocalDate releaseDate, @NonNull int duration) {
-        this.name = name;
-        this.description = description;
-        this.releaseDate = releaseDate;
-        this.duration = duration;
-    }
-}
